Extract public interfaces to abstraction assembly. Extract authorize system as part of mvc project.
