+++ ItspServices.pServer.sln
@@ -20,9 +20,7 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ItspServices.pServer.Persis
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ItspServices.pServer.RepositoryTest", "test\ItspServices.pServer.RepositoryTest\ItspServices.pServer.RepositoryTest.csproj", "{6E4283A0-25AD-45F7-B4AF-8BDDC3BCE285}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ItspServices.pServer.Authorization", "src\ItspServices.pServer.Authorization\ItspServices.pServer.Authorization.csproj", "{43DA547B-2D02-4BDA-9149-04EF5D6FEB3F}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ItspServices.pServer.AuthorizationTest", "test\ItspServices.pServer.AuthorizationTest\ItspServices.pServer.AuthorizationTest.csproj", "{E96EE23D-394C-4AF7-9EC1-F542758827A5}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ItspServices.pServer.AuthorizationTest", "test\ItspServices.pServer.AuthorizationTest\ItspServices.pServer.AuthorizationTest.csproj", "{E96EE23D-394C-4AF7-9EC1-F542758827A5}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -50,10 +48,6 @@ Global
 		{6E4283A0-25AD-45F7-B4AF-8BDDC3BCE285}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{6E4283A0-25AD-45F7-B4AF-8BDDC3BCE285}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{6E4283A0-25AD-45F7-B4AF-8BDDC3BCE285}.Release|Any CPU.Build.0 = Release|Any CPU
-		{43DA547B-2D02-4BDA-9149-04EF5D6FEB3F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{43DA547B-2D02-4BDA-9149-04EF5D6FEB3F}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{43DA547B-2D02-4BDA-9149-04EF5D6FEB3F}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{43DA547B-2D02-4BDA-9149-04EF5D6FEB3F}.Release|Any CPU.Build.0 = Release|Any CPU
 		{E96EE23D-394C-4AF7-9EC1-F542758827A5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{E96EE23D-394C-4AF7-9EC1-F542758827A5}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{E96EE23D-394C-4AF7-9EC1-F542758827A5}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -68,7 +62,6 @@ Global
 		{DC421953-B5A8-4754-85CD-687796DE435B} = {2947A69F-1788-4262-9056-DAA917479403}
 		{2BF99DA8-8E26-4F98-98C8-CCD9C3EAD872} = {2947A69F-1788-4262-9056-DAA917479403}
 		{6E4283A0-25AD-45F7-B4AF-8BDDC3BCE285} = {8F4A9774-BDFD-4039-960D-D750115E290C}
-		{43DA547B-2D02-4BDA-9149-04EF5D6FEB3F} = {2947A69F-1788-4262-9056-DAA917479403}
 		{E96EE23D-394C-4AF7-9EC1-F542758827A5} = {8F4A9774-BDFD-4039-960D-D750115E290C}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
